+++ weblate/urls.py
@@ -309,11 +309,6 @@ real_patterns = [
         name="manage-access",
     ),
     url(
-        r"^access/" + PROJECT + "change/$",
-        weblate.trans.views.acl.change_access,
-        name="change-access",
-    ),
-    url(
         r"^settings/" + PROJECT + "$",
         weblate.trans.views.settings.change_project,
         name="settings",